Moved FastMesh into seperate file
